jekyll-site/requirements.txt
markdownify==0.9.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
